Add getRequestEntriesByEndpointID function
getRequestEntriesByEndpointID is an async function which takes a string (the ID corresponding to a PotatoBin endpoint) and queries MongoDB for matching documents. An array of all documents with the matching endpointID is returned, and an empty array is returned if no documents match the argument passed in.

Rethrowing errors has been added for this function (and the addRequestEntry function in the same file).