Add dbService file to interface between index and DBs
dbService takes responsibility from directly interacting with data layer away from index.js. The addRequest route in index interacts with dbService to update both databases as needed when webhook requests come in.

If the potatoBin being hit does not exist, a 404 is returned, and a 500 is returned if something fails on the data layer. Otherwise, 200 is returned.